+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/openid/DcApiVerifier.kt
@@ -44,7 +44,6 @@ import at.asitplus.wallet.lib.agent.Verifier
 import at.asitplus.wallet.lib.agent.VerifierAgent
 import at.asitplus.wallet.lib.cbor.VerifyCoseSignatureWithKey
 import at.asitplus.wallet.lib.cbor.VerifyCoseSignatureWithKeyFun
-import at.asitplus.wallet.lib.data.CredentialPresentationRequest
 import at.asitplus.wallet.lib.data.CredentialPresentationRequest.DCQLRequest
 import at.asitplus.wallet.lib.data.CredentialPresentationRequest.IsoDeviceRetrieval
 import at.asitplus.wallet.lib.jws.DecryptJwe
@@ -277,29 +276,25 @@ class DcApiVerifier @JvmOverloads constructor(
         Napier.d("validateAuthnResponse: $input")
         val authnRequest = requestFactory.loadAuthnRequest(input, externalId)
 
-        val responseType = authnRequest.responseType?.let { ResponseType.Companion(it) }
+        val expectedNonce = authnRequest.nonce
+            ?: throw IllegalArgumentException("nonce not present in ${authnRequest}")
+        require(nonceAwareVerifier.verifyAndRemoveNonce(expectedNonce)) {
+            "nonce not valid: $expectedNonce, not known to us"
+        }
+
+        val responseType = authnRequest.responseType?.let { ResponseType(it) }
         require(responseType != null) {
             "No response type was specified in the original authentication request."
         }
         require(OpenIdConstants.VP_TOKEN in responseType) {
             "Unsupported response type: $responseType"
         }
-        val expectedNonce = authnRequest.nonce
-            ?: throw IllegalArgumentException("nonce not present in $authnRequest")
-
-        val vpTokenValidationResult = validateVpToken(authnRequest, input, expectedOrigin)
 
         AuthnResponseResult(
             idTokenValidationResult = null,
-            vpTokenValidationResult = vpTokenValidationResult,
+            vpTokenValidationResult = validateVpToken(authnRequest, input, expectedOrigin),
             request = authnRequest,
-        ).also {
-            if (it.isFullyValid()) {
-                require(nonceAwareVerifier.verifyAndRemoveNonce(expectedNonce)) {
-                    "nonce not valid: $expectedNonce, not known to us"
-                }
-            }
-        }
+        )
     }
 
     internal suspend fun validateIsoResponse(
@@ -348,14 +343,6 @@ class DcApiVerifier @JvmOverloads constructor(
 
         Iso180137AnnexCWrapper(documents)
     }
-
-    private fun AuthnResponseResult.isFullyValid(): Boolean =
-        vpTokenValidationResult?.isFailure != true &&
-                (vpTokenValidationResult?.getOrNull()?.isFullyValid() ?: false)
-
-    private fun VpTokenValidationResult.isFullyValid(): Boolean =
-        presentationResults.all { it.isSuccess } &&
-                (this !is VpTokenValidationResultDCQL || submissionRequirementsValidationResult.isSuccess)
 
     /**
      * Validates the `vp_token` of the response with the shared [VpTokenValidator],